Modified the codebase to build under Mac OS X 10.7 SDK with Xcode 4.2.1
This commit modifies the Doom 3 codebase to build under Mac OS X 10.7.  This disables the multi-monitor functionality since it depends on deprecated (and removed from the OS X 10.7 SDK) APIs.  Also modifies the bundled glext.h to build under the new SDK, hopefully without breaking compatibility with older SDKs (although this isn't much of a priority).

Currently there is a bug when linking the "release" version with the templates for interpolations in the UI (the debug version works fine).  To avoid this, the code has been modified to disable interpolations.  This makes the in-game UIs look a big sluggish but the other workaround would be to use the slow debug version.  I might look at this at some point when i find some time, but the major priority was to get a binary for people out ASAP.

Another bug is that the codebase currently doesn't seem to support fullscreen modes properly.  This shouldn't be hard to fix, but again i decided to get out a binary now and try to fix these later.  A custom little program that loads the doom3 binary is provided so that the r_fullscreen var is set to 0 at initialization time.